Avoid zipping when iterating over elements of All proofs
Since the associated types when implementing the `Sigma` trait cannot
be of type `GenericArray`, the type system cannot give us guarantees
that the lists we are working with are of the same length.

With differently sized lists, `zip` will produce an iterator with the
length of the shorter of the two lists, without returning an error or
panicking.

By avoiding `zip` and mapping over the exptected range, we can
therefore fail early instead of producing invalid {announcements,
secret announcements, implied announcements}.