Fix Point<EvenY> serialization/deserialization
Point<EvenY> was serializing to a 33 byte point but only deserialized
from a 32-byte point. Now serialization and Display and deserialization
and FromStr are structured so they match. Debug is now a separate impl
macro rather than being coupled with Display (Debug is a special case in
terms of Points because you need to be able to Debug even Zero points
but not necessarily to Display them or serialize them).