Feature-gate all nightly features behind a `nightly` feature
To better test all combinations of features, we refactor the CI
configuration to use a feature matrix. To ensure that this configuration
doesn't explode in jobs, we switch to the new feature resolver which
allows us to run `cargo test` with `--no-default-features` and
`--features` from the crate root. As such, we can run all tests at once
instead of selecting each individual crate.