Create general neg function for backend
These "any" neg functions work for both normalized and non-normalized
points. Usually, a non-normalized function will work on a normalized
point. Negation is an exception due to how field element normalization
is done in backend. So to support building without specialization we
have to create these very general "any" functions that will always
work (but are a bit slower).