fix: deterministic nonce in secp256k1-sys-0.5.0
secp256k1-sys-0.5.0 introduced breaking changes in its nonce generation
function when no auxiliary data is provided. Now the secret key is alway
bitwise xored with a mask even no auxiliary is provided. The `zero_mask`
used with no auxiliary is the result of:

    TaggedHash("BIP0340/aux", 0x0000..00)

This commit implements the changes to stay compatible with secp-sys in
tests.