Remove xonly in favor of Point<EvenY>
The reason I originally creared XOnly was out of an obsession with not
storing 3 field elements where we could just store 32 bytes.

But if people are in a situation where they don't have enough memory
they can just call `to_xonly_bytes` and use `from_xonly_bytes` when they
want to use it as a key.