Update rest backend with latest k256
This was done by pasting the files over the original and looking at what
changed and then selectively adding those changes.
We're unable to fully update to it because it uses crypto-bigint to do
some scalar operations which we can't as a dependency.
This commit carefully incorporates the state of the upstream as much as possible.
It was extremely tedious work.