Remove *Poly types
We don't want to provide an API for general polynomial arithmetic just
yet. It is very easy to roll yourself anyway. I provide a suggestion
about how to generate scalar polynomials securely that was lacking before.

And other assorted changes to docs and APIs like allow making
verification shares to be zero.