[frost] Make "tweaks" mutate
Prior to this commit tweaks were being stored in a separate field inside
`FrostKey` and were applied at signature combination time. In frostsnap
we actually want to be able to do "one-way" tweaks where you forget the
original polynomial (at least the first coefficient). This is not an API
safety concern but we strictly don't want the original to exist anymore.

The next problem we address is that we don't want you to need the full
polynomial for signing. Instead we make the user provide a "paired"
secret share which is the secret scalar and index paired with the shared
key (first coefficient of polynomial). This is a convenient combination
since it allows you to hash the shared key to figure out a tweak and
then apply it to both the shared key and the shared secret.