[❄] Allow for FROST signature randomization
Creates another session creation method that takes an rng. This ensures
that the final nonce R (and by implication the s) in the signature will
be indistinguishable from random even to all those who knew the
parameters ahead of time. This stops signers from contriving the final
nonce even if they are all corrupt.